backend/requirements.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
autoflake==1.4
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.1
flower==0.9.5
humanize==2.6.0
idna==2.10
isort==5.5.1
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
prometheus-client==0.8.0
ptvsd==4.3.2
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==1.3.0
wrapt==1.12.1
yapf==0.30.0
